requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==2.4.1
attrs==24.2.0
av==15.0.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
decorator==5.1.1
deprecation==2.1.0
executing==2.1.0
faster-whisper==1.2.0
filelock==3.18.0
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2025.7.0
functions-framework==3.8.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.21.0
google-api-python-client==2.178.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.22.0
httptools==0.6.4
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.36.0
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.29.1
networkx==3.4.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
phxsocket==0.1.5
pillow==11.0.0
pillow_heif==0.20.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2025.7.34
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
setuptools==80.9.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
stack-data==0.6.3
streamlit==1.43.1
sympy==1.13.3
tenacity==9.0.0
tifffile==2024.9.20
tokenizers==0.21.4
toml==0.10.2
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.2.3
watchdog==5.0.3
watchfiles==1.1.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
